make compilable under wxWidgets 2.9
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@57770 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/app/cn3d/opengl_renderer.cpp
@@ -947,7 +947,7 @@ void OpenGLRenderer::SetColor(OpenGLRenderer::EColorAction action,
         lastType = cachedType;
     }
 
-    GLfloat rgba[4] = { cr, cg, cb, ca };
+    GLfloat rgba[4] = { (GLfloat) cr, (GLfloat) cg, (GLfloat) cb, (GLfloat) ca };
 #if MAC_GL_OPTIMIZE
 	glMaterialfv(GL_FRONT_AND_BACK, cachedType, rgba);
 #endif
